Started to add AOS4 as build target

+++ src/external/stb_vorbis.c
@@ -645,8 +645,8 @@ typedef unsigned char  uint8;
 typedef   signed char   int8;
 typedef unsigned short uint16;
 typedef   signed short  int16;
-typedef unsigned int   uint32;
-typedef   signed int    int32;
+typedef unsigned long   uint32;
+typedef   signed long    int32;
 
 #ifndef TRUE
 #define TRUE 1
